Move out node graph operations to a separate module
