+++ ayatori/src/graph_representation.rs
@@ -1,10 +1,11 @@
 mod any_node;
 mod args;
 mod constructors;
+mod node_tree;
 mod typed_nodes;
 mod unions;
 
-pub(crate) use any_node::Reproducibility;
+pub(crate) use node_tree::Reproducibility;
 pub(crate) use typed_nodes::{ComputeMappingKind, GeneralizedNode};
 
 #[cfg(feature = "dev")]
